style: jwt 인증 인터셉터 테스트 클래스 하나로 통일 (#272)
